% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveload.R
\name{load_pacmap}
\alias{load_pacmap}
\title{Load a saved PaCMAP model}
\usage{
load_pacmap(file)
}
\arguments{
\item{file}{path to a .rds written by \code{save_pacmap()}.}
}
\description{
Load a saved PaCMAP model
}
